# openapi.json
{
  "components": {
    "schemas": {
      "ApprovalDecisionRequest": {
        "description": "SPEC.md \u00a74.6. API-012: `note` is required (non-empty) when\n`decision` is DENY \u2014 enforced in app/approvals.py, not here, since the\nrule is conditional on the field's own value.",
        "properties": {
          "decision": {
            "enum": [
              "APPROVE",
              "DENY"
            ],
            "title": "Decision",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          }
        },
        "required": [
          "decision"
        ],
        "title": "ApprovalDecisionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Open-source AI Firewall / FWaaS. See SPEC.md.",
    "title": "real-guard-v1",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/console/api/approvals/{approval_id}/decide": {
      "post": {
        "description": "The exact same `app/approvals.decide_and_resume()` call\n`dashboard_decide()` and the canonical\n`POST /v1/firewall/approvals/{id}/decision` both make \u2014 session\nchecked (SEC-005), CSRF checked (SEC-007), `Idempotency-Key` required\n(API-007) \u2014 returning JSON instead of an HTML fragment, because there\nis no HTMX here to swap one into.\n\nSEC-006: this handler never accepts a bearer key at all, only a\nconsole session, and a console session is only ever minted for a\nreviewer identity (`console_login` above). The body is validated\nexplicitly rather than through a FastAPI `body:` parameter so that a\nmalformed body returns *this* module's flat error shape rather than\n`app/main.py`'s transaction-shaped `RequestValidationError` envelope.",
        "operationId": "console_decide_console_api_approvals__approval_id__decide_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Console Decide"
      }
    },
...
    "/dashboard": {
      "get": {
        "description": "SPEC.md \u00a74.1: `GET /dashboard` requires a reviewer session.\nAlso accepts a reviewer bearer key directly (consistency with the\nrest of the API, and useful for `curl`-based verification) \u2014 in\nthat mode the page renders read-only, since there is no session to\nbind a CSRF token to (see docs/adr/0006). A service identity, by\neither route, gets 403. No credential at all redirects to the login\nscreen rather than a bare 401 \u2014 this is a browser-facing page.",
        "operationId": "dashboard_index_dashboard_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "PENDING",
              "title": "Status",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Dashboard Index Dashboard Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dashboard Index"
      }
    },
...
    "/v1/firewall/approvals/{approval_id}/decision": {
      "post": {
        "description": "SPEC.md \u00a74.6. API-012: reviewer-only, Idempotency-Key required,\nnon-empty note required for DENY, self-approval forbidden (APR-009),\nreplay/conflict handling (APR-010).\n\nPhase 4 (ADR 0006): identity resolution now accepts either of\nSPEC.md \u00a74.6's two named auth routes \u2014 `Authorization: Bearer\n<reviewer key>`, or a valid session cookie plus `X-CSRF-Token`\n(SEC-005/SEC-007) \u2014 via `resolve_decision_identity()`, which raises\n`CSRF_TOKEN_INVALID` (403) for a missing/mismatched token on the\ncookie route. In *practice* a real browser never presents this\nendpoint with the SEC-005 cookie automatically, because that\ncookie's `Path=/dashboard` scope means it is only ever attached to\nrequests under `/dashboard` \u2014 this endpoint's cookie support exists\nfor literal SPEC.md \u00a74.6 fidelity and for any client that presents\nthe cookie value explicitly (not from a browser's automatic cookie\njar). The dashboard's own approve/deny buttons instead call\n`POST /dashboard/approvals/{id}/decide` (app/dashboard.py), which\nthe session cookie *does* reach \u2014 see docs/adr/0006 for the full\nreasoning and the concrete proof (a curl-cookie-jar Path test) that\nled to that split.",
        "operationId": "decide_approval_endpoint_v1_firewall_approvals__approval_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Approval Endpoint"
      }
    },
...
  }
}